# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation's  Golang Clean-Architecture",
        "title": "Golang Clean-Architecture",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Chaiyawatkit",
            "url": "https://github.com/chaiyawatkit",
            "email": "chaiyawatkit160340@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all Users on the system",
                "tags": [
                    "user"
                ],
                "summary": "Get all Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.ResponseUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new user on the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create New UserName",
                "parameters": [
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.NewUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    }
                }
            }
        },
        "/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Users by ID on the system",
                "tags": [
                    "user"
                ],
                "summary": "Get users by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "user.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "user.NewUserRequest": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password",
                "role",
                "user"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "Kuk"
                },
                "lastName": {
                    "type": "string",
                    "example": "Mmama"
                },
                "password": {
                    "type": "string",
                    "example": "Password123"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "user": {
                    "type": "string",
                    "example": "Kuk"
                }
            }
        },
        "user.ResponseUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2021-02-24 20:19:39"
                },
                "email": {
                    "type": "string",
                    "example": "kuk@mail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "kuk"
                },
                "id": {
                    "type": "integer",
                    "example": 999
                },
                "lastName": {
                    "type": "string",
                    "example": "Mama"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2021-02-24 20:19:39"
                },
                "user": {
                    "type": "string",
                    "example": "kuk"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}